Bug fix: Fix ReferenceError when clicking the OK or cancel button in "Edit Menu Item"
Bug fix: Fix ReferenceError when clicking the OK or cancel button in the
"Edit Menu Item" dialog (content/edit.xul), on Firefox 45.
content/edit-actions.js declares the variable "EditActions" using "let"
previously, and on Firefox 45, the code in the "ondialogaccept" and
"ondialogcancel" property of the dialog could not access the variable
and throws a ReferenceError, breaking the functionality of the "OK" and
"Cancel" buttons. The commit replaces the "let" with "var" to fix it.